Fix error output and invalid conditions crashing
Up until now, if a configuration condition was invalid
(Wrong syntax, unknown names), it would make Carton crash.
Now we display a passive error and consider the patch empty.

Since a patch should be considered a logical unit, invalidating an
entire one should not be seen as fatal, whereas breaking that logical
unit by, for example, refusing to overwrite a file,  might put the
system in a "partial" state. This is, of course, up for debate,
as no decision is absolute.